+++ src/supadl/transfer/__init__.py
@@ -2,10 +2,18 @@
 
 from supadl.transfer.client import DEFAULT_USER_AGENT, create_http_client
 from supadl.transfer.probe import ProbeResult, ProbeService
+from supadl.transfer.worker import (
+    DEFAULT_TRANSFER_CHUNK_SIZE,
+    SingleStreamWorker,
+    TransferResult,
+)
 
 __all__ = [
+    "DEFAULT_TRANSFER_CHUNK_SIZE",
     "DEFAULT_USER_AGENT",
     "ProbeResult",
     "ProbeService",
+    "SingleStreamWorker",
+    "TransferResult",
     "create_http_client",
 ]